--- app/src/main/java/woowacourse/shopping/view/product/catalog/OnLoadEventListener.kt
@@ -1,5 +1,0 @@
-package woowacourse.shopping.view.product.catalog
-
-fun interface OnLoadEventListener {
-    fun onLoadClick()
-}
