+++ app/src/main/java/woowacourse/shopping/view/product/catalog/allproducts/OnLoadEventListener.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.view.product.catalog.allproducts
+
+fun interface OnLoadEventListener {
+    fun onLoadClick()
+}
